collate decode_tests and split llvm traits
